pyproject.toml
[tool.poetry]
name = "FKQQ_nonebot"
version = "0.1.0"
description = "FKQQ_nonebot"
authors = []
readme = "README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.3"
nonebot-adapter-onebot = "2.1.0"
lxml = "^4.9.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_youthstudy", "nonebot_plugin_petpet", "nonebot_plugin_heweather", "nonebot_plugin_asoulcnki", "nonebot_plugin_logo"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

